Move kidash to a new directory.
This way it can be more easily covnerted into a pip package.
I included some minor changes to the rest to match this change.
