+++ setup.py
@@ -63,10 +63,12 @@ setup(name="grimoire-elk",
         'Topic :: Software Development',
         'License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)',
         'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.4'],
+        'Programming Language :: Python :: 3.4',
+        'Programming Language :: Python :: 3.5'],
       keywords="development repositories analytics",
       packages=['grimoire_elk', 'grimoire_elk.elk', 'grimoire_elk.ocean'],
       scripts=["utils/p2o.py"],
-      install_requires=['perceval>=0.5', 'perceval-mozilla'],
+      install_requires=['perceval>=0.8', 'perceval-mozilla'],
+      include_package_data=True,
       zip_safe=False
     )
